Major changes
Rules rearranged so that the website looks good for browsers without
media queries (IE gets fed the 960px rules separately) and in most
contexts.
Includes a matrix of icons for social networks used as background
images.
Adding class home on the body and home link in the nav so that it can
be hidden when space is scarce.